+++ backend/app/Controllers/Wings/Server/WingsServerInfoController.php
@@ -24,6 +24,7 @@ use App\Chat\Spell;
 use App\Chat\Server;
 use App\Chat\Allocation;
 use App\Chat\ServerVariable;
+use App\Chat\ServerCustomVariable;
 use App\Helpers\ApiResponse;
 use OpenApi\Attributes as OA;
 use App\Plugins\Events\Events\WingsEvent;
@@ -179,6 +180,10 @@ class WingsServerInfoController
         // Build environment variables from server variables
         foreach ($serverVariables as $variable) {
             $environment[$variable['env_variable']] = $variable['variable_value'];
+        }
+
+        foreach (ServerCustomVariable::getEnvironmentVariablesByServerId((int) $server['id']) as $envVariable => $value) {
+            $environment[$envVariable] = $value;
         }
 
         // Add default environment variables based on database fields