+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Plugin sidebar priority system by to control the order of the sidebar items. by @nayskutzu
 - Verify button on the admin user edit page to force verify the user's email by @nayskutzu
 - Verification email resend system in case the user didn't receive the email by @nayskutzu
+- Custom server variables system so users can set variables for their servers and also encrypt them. by @nayskutzu
 
 ### Fixed